+++ .eslintrc.js
@@ -4,6 +4,7 @@ module.exports = {
   rules: {
     'no-plusplus': 'off',
     'import/extensions': 'off',
+    'no-console': [2, { allow: ['warn'] }],
   },
   settings: {
     'import/resolver': {