+++ styleguide/components/navigation/navigation_links_example.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+[
+  {
+    "url" => "/benefits/",
+    "title" => "Benefits"
+  },
+  {
+    "url" => "/work/",
+    "title" => "Work"
+  }
+]